Split files in src and include. Building project with CMake
