+++ include/vectorclock_types.hpp
@@ -1,5 +1,3 @@
 #pragma once
 
-#include "predictor_types.hpp"
-
 typedef int VCValueT;
