+++ src/vectorclock.cpp
@@ -1,4 +1,4 @@
-#include "vectorclock.hpp"
+#include "../include/vectorclock.hpp"
 
 VectorClock::VectorClock() {}
 VectorClock::VectorClock(ThreadIdT increment_thread_id) {
